requirements.txt
Flask
tensorflow
numpy
google-generativeai
gunicorn     # if you deploy
